+++ src/MiningForce/Configuration/ClusterConfig.cs
@@ -115,7 +115,7 @@ namespace MiningForce.Configuration
 	public class PoolPaymentProcessingConfig
 	{
 		public bool Enabled { get; set; }
-		public double MinimumPayment { get; set; }  // in pool-base-currency (ie. Bitcoin, not Satoshis)
+		public decimal MinimumPayment { get; set; }  // in pool-base-currency (ie. Bitcoin, not Satoshis)
 		public PayoutScheme PayoutScheme { get; set; }
 		public JToken PayoutSchemeConfig { get; set; }
 	}
@@ -135,7 +135,7 @@ namespace MiningForce.Configuration
 	{
 		public RewardRecipientType Type { get; set; }
 		public string Address { get; set; }
-		public double Percentage { get; set; }
+		public decimal Percentage { get; set; }
 	}
 
 	public class PoolConfig
