+++ src/MiningForce/Persistence/Repositories/IBalanceRepository.cs
@@ -6,7 +6,7 @@ namespace MiningForce.Persistence.Repositories
 {
     public interface IBalanceRepository
 	{
-		void AddAmount(IDbConnection con, IDbTransaction tx, string poolId, CoinType coin, string address, double amount);
-		Balance[] GetPoolBalancesOverThreshold(IDbConnection con, string poolId, double minimum);
+		void AddAmount(IDbConnection con, IDbTransaction tx, string poolId, CoinType coin, string address, decimal amount);
+		Balance[] GetPoolBalancesOverThreshold(IDbConnection con, string poolId, decimal minimum);
 	}
 }
